refactor: split commands/mod.rs into per-domain submodules
commands/mod.rs had grown to ~1940 lines mixing settings, recording
control, library (snippets/dictionary/app mappings), history,
window/hotkey/autostart, macOS permissions, and the updater into one
file. Split into commands/{settings,recording,library,history,
system,permissions,updater}.rs with mod.rs re-exporting everything,
so tauri::generate_handler! and other call sites are unchanged.

Folded in two small fixes that touched this file: install_update now
resolves its backup path via the shared app_db_path() helper instead
of a separate app.path().app_data_dir() lookup, and set_autostart
quotes the exe path written to the Run registry key so paths with
spaces parse correctly as a single command-line token.
